Clean up tests and add mock
