+++ src/xcode/ENA/ENA/Source/Services/HealthCertificate/Models/__test__/HealthCertifiedPersonTests.swift
@@ -1,0 +1,34 @@
+//
+// 🦠 Corona-Warn-App
+//
+
+import Foundation
+import XCTest
+@testable import ENA
+import OpenCombine
+import HealthCertificateToolkit
+
+class HealthCertifiedPersonTests: XCTestCase {
+
+	func testHealthCertifiedPersonObjectWillChangeTriggered() throws {
+		let healthCertifiedPerson = HealthCertifiedPerson(
+			healthCertificates: [],
+			proofCertificate: nil
+		)
+
+		let objectWillChangeExpectation = expectation(description: "objectWillChange publisher updated")
+
+		let subscription = healthCertifiedPerson.objectWillChange
+			.sink {
+				objectWillChangeExpectation.fulfill()
+			}
+
+		let healthCertificate = HealthCertificate.mock()
+		healthCertifiedPerson.healthCertificates = [healthCertificate]
+
+		waitForExpectations(timeout: 5)
+
+		subscription.cancel()
+	}
+
+}
